requirements.txt
google-generativeai
playwright
moviepy
pyyaml
python-dotenv
apscheduler
playwright-stealth
